[darg] Adds front-to-back password reset capability.
This commit adds complete password reset functionality. A new page is
available on the frontend where you can initiate a password reset
workflow, which sends an e-mail to the input e-mail address containing a
link to a pager where a new password can be entered.

Images representing what the new pages look like can be found in #142.