[auth] Adding authentication middleware and refactoring API
This pushes authentication logic out of our individual route handlers
and into middleware. The authentication middleware adds a `:user` key to
the request map, which is what should be used for authorization by
handlers downstream. This also refactors some of the actual API
endpoints so that they are directly routed to their handler, rather than
to an intermediate authentication handler.