[settings] This adds a settings profile page
This PR adds a settings profile page that lets you change your email and
name. In the future it should also allow you to change your password as
well.

I've only just started thinking about how we'd want to display teams and
team settings, so there's a bit more noodling to do on that. My current
thinking is that you'd have an overall teams setting page with a list of
teams and their settings that you could alter; if it turns out that
people have more than 10 teams then we can design around that but I'm
okay with a simple list for a first pass.

I was thinking that email-sending-times should be configured per-user,
not per-user-team. iDoneThis does per-user-team but honestly I just
don't see that much value add in that. Again, it's a thing that we
could adjust in the future if it turns out to be an issue.