[timezone] Adding a timezone field to users, misc cleanup
This PR does a couple of things. First, it keeps track of a user's
timezone, and provides a dropdown for setting them in the settings page.
It also provides a short-term fix for issue #198 by filling nil values
for the keys (at least of user data in our fixtures) . It also makes it
so that the web command doesn't run `lein clean` first, which seems at
least in part to contribute to our hitting Heroku R10 errors.