[email] Adding a working dropdown for the email time
This adds a working dropdown in a user's profile page for what time of
day they want to get an e-mail from us, and includes logic for when to
send e-mails. It doesn't build the actual e-mail sending component in at
the moment, but the building blocks are there and this branch was
starting to get monstrous.