[fallback.js] Added fallback loader code in HEAD
This commit adds code in the document HEAD to load our JS assets in
order with shims. At the moment it doesn't have CDN links for most of
our dependencies (particularly Angular), but once those are in place we
should be well set for much faster loading times on mobile while still
having support for dev environments and CDN failures.